Understanding and Solving Concurrency issues
Although we use redis watch and transaction to solve the multi concurrent issue,
but the consequence result show alots of failure request occur, which is not the best.
So we use lock and retry mechanism to solve the high failure result issue.
And also need to take care some latency issues.